Concept behind Decision Tree
Decision Trees in Machine Learning:
	Introduction:
CART is one of the first decision tree introduced in Machine Learning for both Classification and Regression problems. Classification and Regression are two kind of problems in Machine Learning. Complexity of Binary tree depends on the height of the Tree. Splitting is one of the key aspects in Decision Tree. 
	For Classification, splitting is done on majority class.
	For regression, splitting is done on average responsible dependent variable of that node. 

	Terminology in Decision Trees:

	Entropy: Entropy is the measure of randomness or unpredictability in the node.
	Information Gain(I(P)): Information Gain is the amount of decrease after the splitting parent node. Information Gain generated by split of parent node P into left child ‘L’ and right child ‘R’ is 
	                            I(P)=G(P)-qG(L)-(1-q)G(R) 
		Where q is (possibly weighted) fraction of instances going towards left.
	Gini Index: 
Gini Index is the name of cost function used to split dataset. Below Gini math formula is used as splitting criteria. 
			G(t)=1-P(t)^2-(1-P(t)^2)
		Where P(t) is possibility weighted of 1 in that node.
	Imbalanced Class: 
While working with large datasets, it is possible to have Null values in attributes. Balancing classes in machine learning is a major issue for practitioners as many data mining methods do not perform well when the training data are highly unbalanced. With CART tree, the practitioner doesn’t have to explicitly balance classes. The CART makes use of ‘priors’ mechanism to provide balance to the tree. 
The math behind prior mechanism is below:
For a binary (0/1) target any node is classified as class 1 if and only if:
       N1(node)/N1(root) >N0(node)/N0(root) 
This allowed CART users to readily use unbalanced data without any data pre-processing to achieve class rebalancing.